src/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
